Add files via upload
Opens a window to select the source path, takes the names of the folders located in the source path, adds a checkbox next to the names, and moves the folders of the selected names into a source path chosen by the user. Useful for when you have an inventory where you create a folder for each product and you need to move the products into a different folder when the product has been sold.

Uses tkinter as gui. Works on Windows. No bugs identified at the moment.